Rs 202 dev

+++ APIs/models/item_db.py
@@ -7,7 +7,7 @@ class ItemModel(db.Model):  # Mapping the class to the rows of the table
     product_id = db.Column(db.Integer, primary_key=True)  # need not be passed
     name = db.Column(db.String(40), unique=True, nullable=False)
     # category = db.Column(db.String(40), unique=True, nullable=False)  # unique to Range->category
-    price = db.Column(db.Integer, unique=False, nullable=False)  # foreign key
+    price = db.Column(db.Float, unique=False, nullable=False)  # foreign key
     store_id = db.Column(db.Integer, db.ForeignKey("stores.store_id"), unique=False, nullable=False)
 
     stores = db.relationship("StoreModel", back_populates="items")
